ACTUALIZACIÓN DE ARCHIVOS

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\UserRequest; 
 use App\Models\User;
 use Illuminate\Http\Request;
+use Spatie\Permission\Models\Role;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class UserController extends Controller
@@ -16,7 +17,7 @@ class UserController extends Controller
     {
         $this->authorize('user-list');
         $texto =$request->input('texto');
-        $registros=User::where('name','like','%'.$texto.'%')
+        $registros=User::with('roles')->where('name','like','%'.$texto.'%')
             ->orWhere('email','like','%'.$texto.'%')
             ->orderBy('id','desc')
             ->paginate(10);
@@ -30,7 +31,8 @@ class UserController extends Controller
     public function create()
     {
         $this->authorize('user-create');
-        return view('usuario.action');
+        $roles=Role::all();
+        return view('usuario.action', compact('roles'));
     }
 
     /**
@@ -45,6 +47,8 @@ class UserController extends Controller
         $registro->password=bcrypt($request->input('password'));
         $registro->activo=$request->input('activo');
         $registro->save();
+
+        $registro->assignRole($request->input('role'));
         return redirect()->route('usuarios.index')->with('mensaje','Registro '.$registro->name. ' agregado correctamente');
     }
 
@@ -62,8 +66,9 @@ class UserController extends Controller
     public function edit($id)
     {
         $this->authorize('user-edit');
+        $roles=Role::all();
         $registro=User::findOrFail($id);
-        return view('usuario.action', compact('registro'));
+        return view('usuario.action', compact('registro','roles'));
     }
 
     /**
@@ -78,6 +83,9 @@ class UserController extends Controller
         $registro->password=bcrypt($request->input('password'));
         $registro->activo=$request->input('activo');
         $registro->save();
+
+    
+
         return redirect()->route('usuarios.index')->with('mensaje','Registro '.$registro->name.' actualizado satisfatoriamente. ' );
     }
 
@@ -86,12 +94,14 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('user-delete');
         $registro=User::findOrFail($id);
         $registro->delete();
         return redirect()->route('usuarios.index')->with('mensaje', $registro->name. ' eliminado satisfatoriamente.');
     }
 
     public function toggleStatus(User $usuario){
+        $this->authorize('user-activate');
         $usuario->activo = !$usuario->activo;
         $usuario->save();
         return redirect()->route('usuarios.index')->with('mensaje', 'El estado del usuario fue actualizado satisfactoriamente');
